Refactoring exists by first and lastName and removing unnecessary code
