--- src/model/service/exists_by_firstname_and_lastname_service.go
@@ -1,5 +1,0 @@
-package service
-
-func (ud *userDomainService) ExistsByFirstNameAndLastName(firstName, lastName string) bool {
-	return ud.repository.ExistsByFirstNameAndLastName(firstName, lastName)
-}
