--- src/model/user_domain_interface.go
@@ -1,19 +1,0 @@
-package model
-
-type UserDomainInterface interface {
-	GetEmail() string
-	GetFirstName() string
-	GetLastName() string
-	GetAge() int8
-	SetId(string)
-	GetId() string
-}
-
-func NewUserDomain(firstName, lastName, email string, age int8) UserDomainInterface {
-	return &userDomain{"",
-		firstName,
-		lastName,
-		email,
-		age,
-	}
-}
