+++ src/repository/user_repository.go
@@ -24,6 +24,18 @@ type userRepository struct {
 	databaseConnection *mongo.Database
 }
 
+func (userRepo *userRepository) ExistsByFirstNameAndLastName(firstName, lastName string) bool {
+	collectionName := viper.GetString(MongoDBUserDb)
+	collection := userRepo.databaseConnection.Collection(collectionName)
+
+	count, _ := collection.CountDocuments(context.Background(), bson.M{"firstName": firstName, "lastName": lastName})
+
+	if count >= 1 {
+		return true
+	}
+	return false
+}
+
 func (userRepo *userRepository) FindUserById(id string) (*models.UserResponse, *custom_errors.CustomErr) {
 	collectionName := viper.GetString(MongoDBUserDb)
 	collection := userRepo.databaseConnection.Collection(collectionName)
