--- src/view/convert_domain_to_response.go
@@ -1,16 +1,0 @@
-package view
-
-import (
-	"github.com/lucasolsi-wex/go-crud/src/model"
-	"github.com/lucasolsi-wex/go-crud/src/models"
-)
-
-func ConvertDomainToResponse(userDomain model.UserDomainInterface) models.UserResponse {
-	return models.UserResponse{
-		Id:        userDomain.GetId(),
-		FirstName: userDomain.GetFirstName(),
-		LastName:  userDomain.GetLastName(),
-		Email:     userDomain.GetEmail(),
-		Age:       userDomain.GetAge(),
-	}
-}
